added menu component
